feat: implement error dto and business base #145
* feat: implement error dto and business base #145

* fix: adjust certificate services #145

* test: fix leadership validation tests #145

* test: use constants to expect entity #145

* fix: remove the error model and remove the second constructor that is in pcts exception #145

* chore: fix formatting of error generation #145

* refactor: add interfaces to make the BusinessBase easier to read

* fix: use new business base #145

* fix(validation): use optional #145

* fix: make function return optional #145

---------

+++ backend/src/main/java/ch/puzzle/pcts/service/business/CertificateTypeBusinessService.java
@@ -1,29 +1,36 @@
 package ch.puzzle.pcts.service.business;
 
+import static ch.puzzle.pcts.Constants.CERTIFICATE_TYPE;
+
+import ch.puzzle.pcts.dto.error.ErrorKey;
+import ch.puzzle.pcts.dto.error.FieldKey;
+import ch.puzzle.pcts.dto.error.GenericErrorDto;
 import ch.puzzle.pcts.exception.PCTSException;
 import ch.puzzle.pcts.model.certificatetype.CertificateType;
-import ch.puzzle.pcts.model.error.ErrorKey;
 import ch.puzzle.pcts.service.persistence.CertificateTypePersistenceService;
 import ch.puzzle.pcts.service.validation.CertificateTypeValidationService;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CertificateTypeBusinessService {
-    private final CertificateTypeValidationService validationService;
-    private final CertificateTypePersistenceService persistenceService;
+public class CertificateTypeBusinessService extends BusinessBase<CertificateType> {
     private final TagBusinessService tagBusinessService;
+    private final CertificateTypePersistenceService certificateTypePersistenceService;
+    private final CertificateTypeValidationService certificateTypeValidationService;
 
     public CertificateTypeBusinessService(CertificateTypeValidationService certificateTypeValidationService,
                                           CertificateTypePersistenceService certificateTypePersistenceService,
                                           TagBusinessService tagBusinessService) {
-        this.validationService = certificateTypeValidationService;
-        this.persistenceService = certificateTypePersistenceService;
+        super(certificateTypeValidationService, certificateTypePersistenceService);
+        this.certificateTypePersistenceService = certificateTypePersistenceService;
+        this.certificateTypeValidationService = certificateTypeValidationService;
         this.tagBusinessService = tagBusinessService;
     }
 
+    @Override
     public CertificateType create(CertificateType certificateType) {
         validationService.validateOnCreate(certificateType);
 
@@ -32,6 +39,7 @@ public class CertificateTypeBusinessService {
         return persistenceService.save(certificateType);
     }
 
+    @Override
     public CertificateType getById(Long id) {
         validationService.validateOnGetById(id);
 
@@ -39,37 +47,69 @@ public class CertificateTypeBusinessService {
 
         if (optionalCertificate.isPresent()) {
             CertificateType certificateType = optionalCertificate.get();
-            validationService.validateCertificateKind(certificateType.getCertificateKind());
+            certificateTypeValidationService.validateCertificateKind(certificateType.getCertificateKind());
             return certificateType;
         } else {
-            throw new PCTSException(HttpStatus.NOT_FOUND,
-                                    "Certificate type with id: " + id + " does not exist.",
-                                    ErrorKey.NOT_FOUND);
+            Map<FieldKey, String> attributes = Map
+                    .of(FieldKey.ENTITY, entityName(), FieldKey.FIELD, "id", FieldKey.IS, id.toString());
+
+            GenericErrorDto error = new GenericErrorDto(ErrorKey.NOT_FOUND, attributes);
+
+            throw new PCTSException(HttpStatus.NOT_FOUND, List.of(error));
+
         }
     }
 
     public List<CertificateType> getAll() {
-        return persistenceService.getAllCertificateTypes();
+        return certificateTypePersistenceService.getAllCertificateTypes();
     }
 
+    @Override
     public CertificateType update(Long id, CertificateType certificateType) {
         validationService.validateOnUpdate(id, certificateType);
 
         certificateType.setTags(tagBusinessService.resolveTags(certificateType.getTags()));
-        certificateType.setId(id);
+        if (persistenceService.getById(id).isPresent()) {
 
-        CertificateType result = persistenceService.save(certificateType);
-        tagBusinessService.deleteUnusedTags();
+            certificateType.setId(id);
 
-        return result;
+            CertificateType result = persistenceService.save(certificateType);
+            tagBusinessService.deleteUnusedTags();
+
+            return result;
+        } else {
+            Map<FieldKey, String> attributes = Map
+                    .of(FieldKey.ENTITY, entityName(), FieldKey.FIELD, "id", FieldKey.IS, id.toString());
+
+            GenericErrorDto error = new GenericErrorDto(ErrorKey.NOT_FOUND, attributes);
+
+            throw new PCTSException(HttpStatus.NOT_FOUND, List.of(error));
+
+        }
     }
 
+    @Override
     public void delete(Long id) {
         validationService.validateOnDelete(id);
 
-        persistenceService.delete(id);
+        if (persistenceService.getById(id).isPresent()) {
+            persistenceService.delete(id);
 
-        tagBusinessService.deleteUnusedTags();
+            tagBusinessService.deleteUnusedTags();
+        } else {
+            Map<FieldKey, String> attributes = Map
+                    .of(FieldKey.ENTITY, entityName(), FieldKey.FIELD, "id", FieldKey.IS, id.toString());
+
+            GenericErrorDto error = new GenericErrorDto(ErrorKey.NOT_FOUND, attributes);
+
+            throw new PCTSException(HttpStatus.NOT_FOUND, List.of(error));
+
+        }
+    }
+
+    @Override
+    protected String entityName() {
+        return CERTIFICATE_TYPE;
     }
 
 }
